third_party/WebKit/LayoutTests/inspector-protocol/css/media-query-listener-exception.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

var theMediaQueryList = window.matchMedia("print");

var callCount = 0;
theMediaQueryList.addListener(function(aMediaQueryList) {
    ++callCount;
    if (callCount == 1) {
        // This needs to use setTimeout because we want to test
        // uncaught exceptions.
        setTimeout("evaluateInFrontend('InspectorTest.resetEmulatedMedia()')", 0);
    }
    if (callCount == 2) {
        // This needs to use setTimeout because we want to test
        // uncaught exceptions.
        setTimeout("evaluateInFrontend('InspectorTest.completeTest()')", 0);
    }
    objectThatDoesNotExist.produceError();
});

function test()
{
    InspectorTest.resetEmulatedMedia = function()
    {
        InspectorTest.sendCommand("Emulation.setEmulatedMedia", { "media": "" });
    }

    function didSetEmulatedMedia(messageObject)
    {
        if (messageObject.error)
            InspectorTest.log("FAILED: " + messageObject.error.message);
    }

    InspectorTest.sendCommand("Emulation.setEmulatedMedia", {
        "media": "print"
    }, didSetEmulatedMedia);
}
</script>
</head>
<body onload="runTest()">
<p>Test that uncaught exception in MediaQueryListListener will be reported to the console. On success you should see two exceptions in the listener logged to the console (first time when the media type is overridden and second - when they are restored). <a href="https://bugs.webkit.org/show_bug.cgi?id=105162">Bug 105162.</p>
</body>
</html>
