third_party/blink/web_tests/inspector-protocol/css/media-query-listener-exception-expected.txt

Test that uncaught exception in MediaQueryListListener will be reported to the console. On success you should see two exceptions in the listener logged to the console (first time when the media type is overridden and second - when they are restored). https://bugs.webkit.org/show_bug.cgi?id=105162
ReferenceError: objectThatDoesNotExist is not defined
    at MediaQueryList.<anonymous> (<anonymous>:3:64)
ReferenceError: objectThatDoesNotExist is not defined
    at MediaQueryList.<anonymous> (<anonymous>:3:64)
